Moving component render to App.tsx
